# api.yaml
openapi: 3.0.3
info:
  title: Art Explorer API
  description: API para la aplicación de exploración de arte, gestión de colecciones de usuarios y catálogo de obras.
  version: 1.0.0
servers:
  - url: https://api.artexplorer.com
    description: Servidor de Producción
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

    webhookSecret:
      type: apiKey
      in: header
      name: x-webhook-secret
      description: Secret key for validating Supabase webhook requests
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: 'Invalid image format'
        message:
          type: string
          example: 'The uploaded file must be a valid JPG, PNG, or HEIC image'
...
    CreatePaymentIntentRequest:
      type: object
      description: 'Request body para crear Payment Intent de Travel Pass'
      required:
        - planId
      properties:
        planId:
          type: string
          enum: [travel_pass]
          description: 'Identificador del plan - solo se acepta travel_pass para payment intents'
          example: 'travel_pass'
        currency:
          type: string
          default: 'eur'
          description: 'Código de moneda (ISO 4217)'
          example: 'eur'
    PaymentSheetResponse:
      type: object
      description: 'Credenciales para inicializar Stripe Payment Sheet en el cliente'
      required:
        - paymentIntent
        - ephemeralKey
        - customer
        - publishableKey
      properties:
        paymentIntent:
          type: string
          description: 'Payment Intent client_secret (comienza con pi_)'
          example: 'pi_3QLxyz123_secret_xyz789abc'
        ephemeralKey:
          type: string
          description: 'Ephemeral Key secret para autenticación del Stripe SDK'
          example: 'ek_test_YWJjZGVm'
        customer:
          type: string
          description: 'Stripe Customer ID (comienza con cus_)'
          example: 'cus_QtPxyz123ABC'
        publishableKey:
          type: string
          description: 'Stripe publishable key para inicializar el SDK'
          example: 'pk_test_51xyz...'
    CreateSetupIntentRequest:
      type: object
      description: 'Request body para crear SetupIntent de suscripción (Paso 1/2)'
      required:
        - planId
      properties:
        planId:
          type: string
          enum: [monthly, annual]
          description: 'Identificador del plan de suscripción'
          example: 'monthly'
    SetupIntentResponse:
      type: object
      description: 'Credenciales para inicializar Stripe SetupIntent Sheet en el cliente'
      required:
        - setupIntent
        - ephemeralKey
        - customer
        - publishableKey
      properties:
        setupIntent:
          type: string
          description: 'SetupIntent client_secret (comienza con seti_)'
          example: 'seti_1QLxyz123_secret_abc789xyz'
        ephemeralKey:
          type: string
          description: 'Ephemeral Key secret para autenticación del Stripe SDK'
          example: 'ek_test_YWJjZGVm'
        customer:
          type: string
          description: 'Stripe Customer ID (comienza con cus_)'
          example: 'cus_QtPxyz123ABC'
        publishableKey:
          type: string
          description: 'Stripe publishable key para inicializar el SDK'
          example: 'pk_test_51xyz...'
    CreateSubscriptionRequest:
      type: object
      description: 'Request body para crear suscripción con método de pago guardado (Paso 2/2)'
      required:
        - planId
      properties:
        planId:
          type: string
          enum: [monthly, annual]
          description: 'Identificador del plan de suscripción'
          example: 'monthly'
        setupIntentId:
          type: string
          pattern: '^seti_[a-zA-Z0-9_]+$'
          description: 'Opcional: SetupIntent ID para extraer método de pago (comienza con seti_)'
          example: 'seti_1QLxyz123abc'
    SubscriptionCreatedResponse:
      type: object
      description: 'Respuesta al crear una suscripción exitosamente'
      required:
        - subscriptionId
        - status
        - currentPeriodEnd
      properties:
        subscriptionId:
          type: string
          description: 'Stripe Subscription ID (comienza con sub_)'
          example: 'sub_1QLxyzABC123'
        status:
          type: string
          description: 'Estado actual de la suscripción'
          example: 'active'
        currentPeriodEnd:
          type: string
          format: date-time
          description: 'Fecha ISO 8601 cuando termina el período de facturación actual'
          example: '2026-01-15T14:30:00.000Z'
...
paths:
...
  /api/payments/create-intent:
    post:
      tags:
        - Payments
      summary: Crear Payment Intent para Travel Pass
      description: |
        Crea un Stripe Payment Intent para la compra única del Travel Pass.

        **Detalles del Travel Pass:**
        - Precio: €4.99 (pago único)
        - Duración: 7 días de acceso premium
        - Sin renovación automática

        **Flujo de Pago:**
        1. Cliente llama a este endpoint para crear Payment Intent
        2. Backend retorna credenciales para Stripe Payment Sheet
        3. Cliente inicializa Stripe SDK con credenciales retornadas
        4. Usuario completa el pago en Stripe Payment Sheet
        5. Webhook procesa la confirmación del pago
        6. Usuario obtiene 7 días de acceso premium

        **Reglas de Negocio:**
        - Usuario no puede tener suscripción o travel pass activo existente
        - Bloquea compras duplicadas mientras el usuario tiene premium activo
        - Crea o recupera cliente de Stripe automáticamente
        - Confirmación de pago manejada vía webhook (payment_intent.succeeded)

        **Rate Limiting:**
        - Se aplican límites estándar de la API
        - Considerar implementar rate limiting adicional para endpoints de pago
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentIntentRequest'
            examples:
              travel_pass:
                summary: Crear payment intent de Travel Pass
                value:
                  planId: 'travel_pass'
                  currency: 'eur'
      responses:
        '200':
          description: Payment Intent creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentSheetResponse'
              example:
                paymentIntent: 'pi_3QLxyz123_secret_xyz789abc'
                ephemeralKey: 'ek_test_YWJjZGVm'
                customer: 'cus_QtPxyz123ABC'
                publishableKey: 'pk_test_51xyz...'
        '400':
          description: Usuario ya tiene premium activo o solicitud inválida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                already_premium:
                  summary: Usuario ya tiene suscripción o travel pass activo
                  value:
                    success: false
                    error: 'bad_request'
                    message: 'Ya tienes una suscripción o pase activo. Solo puedes tener una suscripción activa a la vez.'
                invalid_plan:
                  summary: PlanId inválido proporcionado
                  value:
                    success: false
                    error: 'validation_error'
                    message: 'planId must be travel_pass for payment intents'
        '401':
          description: No autenticado - token JWT faltante o inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: 'unauthorized'
                message: 'Valid authentication token required'
        '500':
          description: Error interno del servidor - fallo en API de Stripe o base de datos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: 'internal_server_error'
                message: 'Failed to create payment intent'
  /api/payments/setup-intent:
    post:
      tags:
        - Payments
      summary: Crear SetupIntent para Suscripción (Paso 1 de 2)
      description: |
        Crea un Stripe SetupIntent para recolectar método de pago sin cobrar al usuario.
        Este es el **Paso 1** del flujo de suscripción en 2 pasos.

        **Precios de Suscripción:**
        - Plan Mensual: €8.99/mes (recurrente)
        - Plan Anual: €59.99/año (recurrente)

        **Flujo de Suscripción en 2 Pasos:**

        **Paso 1 (este endpoint)**: Recolectar método de pago
        1. Cliente llama a este endpoint con el plan deseado
        2. Backend crea SetupIntent (sin cobro)
        3. Cliente inicializa Stripe SetupIntent Sheet
        4. Usuario ingresa detalles de la tarjeta
        5. Stripe guarda el método de pago
        6. Cliente recibe setupIntentId

        **Paso 2**: Crear suscripción con método de pago guardado
        - Cliente llama a `/api/payments/create-subscription-with-payment-method`
        - Backend crea suscripción con cobro inmediato
        - Usuario obtiene acceso premium

        **Reglas de Negocio:**
        - Usuario no puede tener suscripción o travel pass activo existente
        - SetupIntent solo guarda método de pago - NO cobra
        - Método de pago puede usarse después para crear suscripción
        - SetupIntent tiene usage='off_session' para cobros recurrentes

        **Por qué Flujo en 2 Pasos:**
        - Cumple con requisitos de Strong Customer Authentication (SCA)
        - Permite al usuario revisar detalles de suscripción antes del cobro
        - Proporciona mejor manejo de errores para validación de método de pago
        - Habilita creación fluida de suscripción después de confirmación
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSetupIntentRequest'
            examples:
              monthly_plan:
                summary: Configurar método de pago para suscripción mensual
                value:
                  planId: 'monthly'
              annual_plan:
                summary: Configurar método de pago para suscripción anual
                value:
                  planId: 'annual'
      responses:
        '200':
          description: SetupIntent creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SetupIntentResponse'
              example:
                setupIntent: 'seti_1QLxyz123_secret_abc789xyz'
                ephemeralKey: 'ek_test_YWJjZGVm'
                customer: 'cus_QtPxyz123ABC'
                publishableKey: 'pk_test_51xyz...'
        '400':
          description: Usuario ya tiene premium activo o solicitud inválida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                already_premium:
                  summary: Usuario ya tiene suscripción o travel pass activo
                  value:
                    success: false
                    error: 'bad_request'
                    message: 'Ya tienes una suscripción o pase activo. Solo puedes tener una suscripción activa a la vez.'
                invalid_plan:
                  summary: PlanId inválido proporcionado
                  value:
                    success: false
                    error: 'validation_error'
                    message: "planId must be 'monthly' or 'annual'"
        '401':
          description: No autenticado - token JWT faltante o inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: 'unauthorized'
                message: 'Valid authentication token required'
        '500':
          description: Error interno del servidor - fallo en API de Stripe o base de datos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: 'internal_server_error'
                message: 'Failed to create setup intent'
  /api/payments/create-subscription-with-payment-method:
    post:
      tags:
        - Payments
      summary: Crear Suscripción con Método de Pago (Paso 2 de 2)
      description: |
        Crea una Suscripción de Stripe usando un método de pago guardado.
        Este es el **Paso 2** del flujo de suscripción en 2 pasos.

        **Prerequisitos:**
        - Usuario debe haber completado Paso 1 (POST /api/payments/setup-intent)
        - SetupIntent debe estar en estado 'succeeded'
        - Usuario no debe tener suscripción activa existente

        **Resolución de Método de Pago:**
        1. **Con setupIntentId** (recomendado): Extrae método de pago del SetupIntent completado
        2. **Sin setupIntentId**: Usa método de pago por defecto del cliente o primera tarjeta adjunta

        **Precios de Suscripción:**
        - Plan Mensual: €8.99/mes (cobrado inmediatamente, luego recurrente mensual)
        - Plan Anual: €59.99/año (cobrado inmediatamente, luego recurrente anual)

        **Lógica de Negocio:**
        1. Valida que usuario no tenga suscripción activa
        2. Recupera/valida método de pago
        3. Mapea planId a Stripe Price ID desde configuración de entorno
        4. Crea Suscripción de Stripe (cobro inmediato)
        5. Almacena suscripción en base de datos
        6. Actualiza user.isPremium = true (si suscripción activa)
        7. Retorna detalles de suscripción

        **Webhooks:**
        - `customer.subscription.created`: Confirma creación de suscripción
        - `invoice.paid`: Confirma pago inicial y renovaciones futuras
        - `invoice.payment_failed`: Maneja pagos fallidos

        **Ciclo de Vida de Suscripción:**
        - Active: Usuario tiene acceso premium
        - Past Due: Pago falló, período de gracia activo
        - Canceled: Usuario o sistema canceló suscripción
        - Expired: Suscripción terminó (sin renovación)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubscriptionRequest'
            examples:
              with_setup_intent:
                summary: Crear suscripción con SetupIntent (recomendado)
                value:
                  planId: 'monthly'
                  setupIntentId: 'seti_1QLxyz123abc'
              without_setup_intent:
                summary: Crear suscripción con método de pago por defecto del cliente
                value:
                  planId: 'annual'
      responses:
        '200':
          description: Suscripción creada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionCreatedResponse'
              examples:
                active_subscription:
                  summary: Suscripción creada y activa
                  value:
                    subscriptionId: 'sub_1QLxyzABC123'
                    status: 'active'
                    currentPeriodEnd: '2026-01-15T14:30:00.000Z'
                incomplete_subscription:
                  summary: Suscripción creada pero pago pendiente
                  value:
                    subscriptionId: 'sub_1QLxyzABC456'
                    status: 'incomplete'
                    currentPeriodEnd: '2026-01-15T14:30:00.000Z'
        '400':
          description: Solicitud inválida o violación de regla de negocio
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                already_subscribed:
                  summary: Usuario ya tiene suscripción activa
                  value:
                    success: false
                    error: 'bad_request'
                    message: 'Ya tienes una suscripción activa'
                setup_intent_not_completed:
                  summary: SetupIntent aún no completado por usuario
                  value:
                    success: false
                    error: 'bad_request'
                    message: 'SetupIntent no completado. Estado: requires_payment_method'
                no_payment_method:
                  summary: No se encontró método de pago para el cliente
                  value:
                    success: false
                    error: 'bad_request'
                    message: 'No se encontró ningún método de pago para este cliente.'
                payment_method_mismatch:
                  summary: Método de pago pertenece a cliente diferente
                  value:
                    success: false
                    error: 'bad_request'
                    message: 'El método de pago no pertenece a este cliente'
                price_not_configured:
                  summary: Price ID no configurado en entorno
                  value:
                    success: false
                    error: 'bad_request'
                    message: 'Price ID not configured for plan: monthly'
        '401':
          description: No autenticado - token JWT faltante o inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: 'unauthorized'
                message: 'Valid authentication token required'
        '500':
          description: Error interno del servidor - fallo en API de Stripe o base de datos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: 'internal_server_error'
                message: 'Failed to create subscription'
  /api/payments/subscription/status:
    get:
      tags:
        - Payments
      summary: Obtener Estado de Suscripción del Usuario
      description: |
        Retorna el estado actual de suscripción o travel pass para el usuario autenticado.

        **Prioridad de Estado:**
        1. **Compra Única (Travel Pass)**: Verificada primero
           - Retorna si usuario tiene travel pass activo (validUntil > now, status=succeeded)
           - Retorna: `plan: null, status: 'active', isSubscribed: true`

        2. **Suscripción**: Verificada segundo
           - Retorna si usuario tiene suscripción con estado 'active' o 'past_due'
           - Auto-expira si currentPeriodEnd < now (establece status='expired', isPremium=false)
           - Retorna: `plan: 'monthly'|'annual', status: 'active'|'past_due'|'expired'`

        3. **Sin Premium**: Sin compra o suscripción activa
           - Retorna: `plan: null, status: 'none', isSubscribed: false`

        **Valores de Estado:**
        - `active`: Usuario tiene acceso premium activo
        - `past_due`: Pago de suscripción falló, período de gracia activo
        - `canceled`: Suscripción cancelada, acceso puede continuar hasta fin de período
        - `expired`: Suscripción o travel pass expirado
        - `none`: Sin suscripción o travel pass

        **Casos de Uso:**
        - Verificar si usuario necesita renovar premium
        - Mostrar detalles de suscripción en configuración
        - Hacer cumplir acceso a funciones premium
        - Mostrar prompts de actualización apropiados

        **Auto-Expiración:**
        - Si subscription.currentPeriodEnd < now: estado establecido a 'expired', isPremium=false
        - Travel pass auto-expira basado en fecha validUntil
        - Cliente debe llamar periódicamente a este endpoint para sincronizar estado
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Estado de suscripción obtenido exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionStatusResponse'
              examples:
                active_travel_pass:
                  summary: Usuario tiene Travel Pass activo
                  value:
                    plan: null
                    status: 'active'
                    expiryDate: '2025-12-23T14:30:00.000Z'
                    isSubscribed: true
                active_monthly_subscription:
                  summary: Usuario tiene suscripción mensual activa
                  value:
                    plan: 'monthly'
                    status: 'active'
                    expiryDate: '2026-01-15T14:30:00.000Z'
                    isSubscribed: true
                active_annual_subscription:
                  summary: Usuario tiene suscripción anual activa
                  value:
                    plan: 'annual'
                    status: 'active'
                    expiryDate: '2026-12-15T14:30:00.000Z'
                    isSubscribed: true
                past_due_subscription:
                  summary: Pago de suscripción falló (período de gracia)
                  value:
                    plan: 'monthly'
                    status: 'past_due'
                    expiryDate: '2025-12-15T14:30:00.000Z'
                    isSubscribed: false
                expired_subscription:
                  summary: Suscripción expirada
                  value:
                    plan: 'monthly'
                    status: 'expired'
                    expiryDate: '2025-11-15T14:30:00.000Z'
                    isSubscribed: false
                no_subscription:
                  summary: Usuario no tiene suscripción ni travel pass
                  value:
                    plan: null
                    status: 'none'
                    isSubscribed: false
        '401':
          description: No autenticado - token JWT faltante o inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: 'unauthorized'
                message: 'Valid authentication token required'
        '500':
          description: Error interno del servidor - fallo en base de datos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: 'internal_server_error'
                message: 'Failed to retrieve subscription status'
  /api/payments/subscription/cancel:
    post:
      tags:
        - Payments
      summary: Cancelar Suscripción al Final del Período
      description: |
        Cancela la suscripción activa del usuario al final del período de facturación actual.

        **Comportamiento:**
        - Marca la suscripción para cancelación en Stripe con `cancel_at_period_end: true`
        - Usuario mantiene acceso premium hasta `currentPeriodEnd`
        - No se emiten reembolsos (política de no reembolso)
        - No se realizarán más cargos después del período actual
        - Webhook `customer.subscription.deleted` se dispara en la fecha de finalización

        **Restricciones:**
        - Solo suscripciones mensuales y anuales pueden ser canceladas
        - Travel Pass (compra única) NO puede ser cancelado ni reembolsado
        - No se puede cancelar una suscripción ya cancelada
        - Usuario debe tener suscripción activa o past_due

        **Flujo de Cancelación:**
        1. Cliente llama a este endpoint
        2. Backend marca suscripción con `cancel_at_period_end: true` en Stripe
        3. Usuario continúa con acceso premium hasta `currentPeriodEnd`
        4. En fecha de expiración, Stripe dispara webhook `customer.subscription.deleted`
        5. Webhook establece `status: 'canceled'` y `isPremium: false`

        **Después de Cancelación:**
        - Estado de suscripción permanece 'active' hasta el fin del período
        - Para reactivar, usuario debe crear nueva suscripción (no hay reactivación)
        - `cancel_at_period_end` se puede verificar en respuesta de este endpoint

        **Casos de Uso:**
        - Usuario desea cancelar antes de la próxima renovación
        - Usuario probó premium y no desea continuar
        - Usuario cambia a diferentes opciones de pago/plan (cancelar primero)
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              description: No se requieren parámetros en el body
            example: {}
      responses:
        '200':
          description: Suscripción marcada para cancelación exitosamente
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - subscription
                  - message
                properties:
                  success:
                    type: boolean
                    description: Indica si la cancelación fue exitosa
                    example: true
                  subscription:
                    type: object
                    required:
                      - id
                      - status
                      - cancelAtPeriodEnd
                      - currentPeriodEnd
                      - planId
                    properties:
                      id:
                        type: string
                        description: ID de suscripción de Stripe
                        example: 'sub_1QLxyzABC123'
                      status:
                        type: string
                        description: Estado actual de la suscripción (permanece 'active' hasta el fin del período)
                        enum: [active, past_due]
                        example: 'active'
                      cancelAtPeriodEnd:
                        type: boolean
                        description: Indica si la suscripción está marcada para cancelación
                        example: true
                      currentPeriodEnd:
                        type: string
                        format: date-time
                        description: Fecha cuando la suscripción será cancelada (formato ISO 8601)
                        example: '2026-01-15T14:30:00.000Z'
                      planId:
                        type: string
                        description: Tipo de plan de suscripción
                        enum: [monthly, annual]
                        example: 'monthly'
                  message:
                    type: string
                    description: Mensaje descriptivo sobre la cancelación
                    example: 'Tu suscripción será cancelada el 15 de enero de 2026. Seguirás teniendo acceso premium hasta esa fecha.'
              examples:
                monthly_subscription_canceled:
                  summary: Suscripción mensual cancelada exitosamente
                  value:
                    success: true
                    subscription:
                      id: 'sub_1QLxyzABC123'
                      status: 'active'
                      cancelAtPeriodEnd: true
                      currentPeriodEnd: '2026-01-15T14:30:00.000Z'
                      planId: 'monthly'
                    message: 'Tu suscripción será cancelada el 15 de enero de 2026. Seguirás teniendo acceso premium hasta esa fecha.'
                annual_subscription_canceled:
                  summary: Suscripción anual cancelada exitosamente
                  value:
                    success: true
                    subscription:
                      id: 'sub_1QLxyzABC456'
                      status: 'active'
                      cancelAtPeriodEnd: true
                      currentPeriodEnd: '2026-12-15T14:30:00.000Z'
                      planId: 'annual'
                    message: 'Tu suscripción será cancelada el 15 de diciembre de 2026. Seguirás teniendo acceso premium hasta esa fecha.'
        '400':
          description: Solicitud inválida - sin suscripción, ya cancelada, o solo Travel Pass
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                no_subscription:
                  summary: Usuario no tiene suscripción activa
                  value:
                    success: false
                    error: 'not_found'
                    message: 'No tienes una suscripción activa'
                already_canceled:
                  summary: Suscripción ya está cancelada
                  value:
                    success: false
                    error: 'bad_request'
                    message: 'La suscripción ya está cancelada'
                travel_pass_only:
                  summary: Usuario solo tiene Travel Pass (no cancelable)
                  value:
                    success: false
                    error: 'bad_request'
                    message: 'No tienes una suscripción activa. Los Travel Pass no pueden ser cancelados.'
                stripe_error:
                  summary: Error al comunicarse con Stripe
                  value:
                    success: false
                    error: 'bad_request'
                    message: 'Error al cancelar la suscripción. Por favor, intenta nuevamente.'
        '401':
          description: No autenticado - token JWT faltante o inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: 'unauthorized'
                message: 'Valid authentication token required'
        '404':
          description: Suscripción no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: 'not_found'
                message: 'No tienes una suscripción activa'
        '500':
          description: Error interno del servidor - fallo en API de Stripe o base de datos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: 'internal_server_error'
                message: 'Error al cancelar la suscripción. Por favor, intenta nuevamente.'
...